+++ src/common/Navbar.jsx
@@ -1,4 +1,4 @@
-import { Offcanvas, Button } from 'react-bootstrap';
+import { Button } from 'react-bootstrap';
 import { useEffect, useState } from "react";
 import { useNavigate, Link } from 'react-router-dom';
 import { signOut, testAuth } from "../api/signApi";
@@ -8,6 +8,7 @@ import 'bootstrap-icons/font/bootstrap-icons.css';
 import UserAuthentication from '../sign/service/UserAuthentication';
 import { toast } from 'react-toastify';
 import useAlert from '../hooks/useAlert';
+import NavMenu from './NavMenu';
 
 const GRADIENT = "linear-gradient(90deg, #5C6BC0 0%, #283593 100%)";
 
@@ -69,9 +70,7 @@ const Navbar = () => {
 
   return (
     <>
-      <div
-        style={{ position: "fixed", top: 80, left: "50%", transform: "translateX(-50%)", minWidth: 260, zIndex: 2000 }}
-      >
+      <div style={{ position: "fixed", top: 80, left: "50%", transform: "translateX(-50%)", minWidth: 260, zIndex: 2000 }}>
         {alert.show && (
           <div className={`alert alert-${alert.type || "info"} text-center shadow`} role="alert">
             {alert.message}
@@ -107,71 +106,19 @@ const Navbar = () => {
             </Button>
           </div>
         </div>
-        <Offcanvas show={show} onHide={handleClose} placement="end" id="offcanvasNavbar"
-          aria-labelledby="offcanvasNavbarLabel"
-          style={{
-            width: '300px',
-            background: "linear-gradient(135deg, #283593D9 60%, #6F8AE7DD 100%)",
-            color: "white", borderTopLeftRadius: "28px", borderBottomLeftRadius: "28px",
-            boxShadow: "0 0 32px 0 rgba(40,53,147,0.13)",
-            backdropFilter: "blur(10px)", WebkitBackdropFilter: "blur(10px)"
-          }}>
-          <Offcanvas.Header closeButton
-            style={{ background: "none", color: "white", borderBottom: "1px solid rgba(255,255,255,0.10)" }}>
-            <Offcanvas.Title id="offcanvasNavbarLabel"
-              style={{ fontWeight: 600, letterSpacing: "0.05em", fontSize: "1.2rem" }}>
-              <i className="bi bi-menu-button-wide me-2" />메뉴
-            </Offcanvas.Title>
-          </Offcanvas.Header>
-          <Offcanvas.Body>
-            <div className="d-flex flex-column justify-content-between h-100">
-              <div className="d-flex flex-column align-items-stretch gap-2">
-                <MenuLink to="/post/list" icon="bi-card-list" label="여행지" onClick={handleClose} />
-                {!isLoggedIn && (
-                  <>
-                    <MenuLink to="sign/component/signinpage" icon="bi-box-arrow-in-right" label="로그인"
-                      onClick={() => { goToLogin(); handleClose(); }} />
-                    <MenuLink to="sign/component/signuppage" icon="bi-person-plus" label="회원 가입"
-                      onClick={() => { goToSignup(); handleClose(); }} />
-                  </>
-                )}
-                {isLoggedIn && (
-                  <>
-                    {isAdmin &&
-                      <MenuLink to="/component/admnpage" icon="bi-gear" label="관리자 메뉴" onClick={handleClose} />}
-                    <MenuLink to="/common/MyPage" icon="bi-person-circle" label="마이페이지" onClick={handleClose} />
-                  </>
-                )}
-              </div>
-              <div className="d-flex flex-column align-items-stretch gap-2 mt-4">
-                {isLoggedIn && <MenuLink to="#" icon="bi-box-arrow-right" label="로그아웃" onClick={handleLogout} />}
-                {isLoggedIn &&
-                  <span className="nav-link w-100 fs-6 text-light bg-secondary bg-opacity-50 rounded px-3 py-2 disabled"
-                    style={{ pointerEvents: 'none', opacity: 0.7, background: "rgba(100, 100, 180, 0.16)" }}>
-                    <i className="bi bi-person-badge me-2"></i>닉네임: {curUser}
-                  </span>
-                }
-              </div>
-            </div>
-          </Offcanvas.Body>
-        </Offcanvas>
+        <NavMenu
+          show={show}
+          onClose={handleClose}
+          isLoggedIn={isLoggedIn}
+          isAdmin={isAdmin}
+          curUser={curUser}
+          onLogout={handleLogout}
+          onLogin={() => { goToLogin(); handleClose(); }}
+          onSignup={() => { goToSignup(); handleClose(); }}
+        />
       </nav>
     </>
   );
 };
 
 export default Navbar;
-
-const MenuLink = ({ to, icon, label, onClick, fs = "fs-5" }) => (
-  <Link
-    to={to}
-    className={`nav-link w-100 ${fs} text-light bg-opacity-75 rounded px-3 py-2`}
-    style={{ position: "relative", transition: "background 0.2s, box-shadow 0.16s", fontWeight: 500, borderRadius: "18px", letterSpacing: "0.01em" }}
-    onClick={onClick}
-    onMouseOver={e => { e.target.style.background = "rgba(255,255,255,0.11)"; e.target.style.color = "#FFF"; e.target.style.boxShadow = "0 2px 12px 0 rgba(91,142,255,0.09)"; }}
-    onMouseOut={e => { e.target.style.background = ""; e.target.style.color = "#FFF"; e.target.style.boxShadow = ""; }}
-  >
-    <i className={`bi ${icon} me-2`} style={{ opacity: 0.96 }} />
-    {label}
-  </Link>
-);
