+++ src/post/pages/PostEditPage.jsx
@@ -1,11 +1,10 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.bundle.min.js';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import RadioPage from '../../board/component/page/RegionRadioComp';
 import { useEffect, useState } from 'react';
 import { getPost, editPost, removePost } from '../../api/postApi';
-import CategoryCard from '../../board/component/page/CategoryCard';
 import useAlert from '../../hooks/useAlert';
+import PostForm from '../components/PostForm';
 
 const PostEditPage = () => {
     const [searchParams] = useSearchParams();
@@ -22,7 +21,6 @@ const PostEditPage = () => {
         region: '',
         imagePaths: []
     });
-
     const [existingImages, setExistingImages] = useState([]);
     const [newImages, setNewImages] = useState([]);
     const [imagePreviews, setImagePreviews] = useState([]);
@@ -52,7 +50,7 @@ const PostEditPage = () => {
 
     useEffect(() => {
         viewBoard();
-        let nickname = localStorage.getItem("nickname");
+        const nickname = localStorage.getItem("nickname");
         if (nickname == null) {
             showAlert("로그인이 필요합니다.", "danger");
             setTimeout(() => navigate(-1), 500);
@@ -114,117 +112,59 @@ const PostEditPage = () => {
         }
     };
 
-    return (
-        <>
-            {alert.show && (
-                <div className={`alert alert-${alert.type} text-center shadow`} role="alert"
-                    style={{ position: "fixed", top: 80, left: "50%", transform: "translateX(-50%)", minWidth: 240, zIndex: 3000 }}>
-                    {alert.message}
-                </div>
-            )}
-            <div className="py-5" style={{ minHeight: "100vh", width: "100vw", overflowX: "hidden", position: "relative" }}>
-                <div className="container my-5" style={{ maxWidth: '900px' }}>
-                    <div className="card shadow-lg border-0 rounded-4 p-4" style={{ background: "#ffffffeb" }}>
-                        <h2 className="mb-3 fw-bold" style={{ textAlign: 'center', letterSpacing: '2px' }}>글 수정</h2>
-                        <div className="text-secondary text-center mb-4" style={{ fontSize: '1.07rem' }}>
-                            여행지, 사진, 지역, 카테고리, 후기를 모두 입력해 주세요!
+    const imageSection = (
+        <div className="mb-4">
+            <label className="form-label fw-semibold">
+                사진 첨부 <span className="text-secondary" style={{ fontSize: "0.95em" }}>(여러 장 첨부 가능)</span>
+            </label>
+            <div className="bg-light rounded-4 p-3 px-4 border">
+                <input type="file" accept="image/*" multiple onChange={handleImageChange} className="form-control mb-3" />
+                <div className="d-flex flex-wrap gap-3">
+                    {existingImages.map((src, idx) => (
+                        <div key={`exist-${idx}`} style={{ position: 'relative' }}>
+                            <img
+                                src={src.startsWith('/images/') ? `${process.env.REACT_APP_IMAGE_BASE_URL}${src}` : src}
+                                alt={`preview-exist-${idx}`}
+                                style={{ width: 100, height: 100, objectFit: 'cover', borderRadius: 14, border: '1px solid #eee', boxShadow: "0 2px 6px rgba(0,0,0,0.06)" }}
+                            />
+                            <button
+                                type="button" className="btn btn-danger btn-sm"
+                                style={{ position: 'absolute', top: 5, right: 5, borderRadius: '50%', padding: '2px 7px', fontSize: "1.05rem" }}
+                                onClick={() => handleExistingImageRemove(idx)}
+                            >×</button>
                         </div>
-                        <div className="mb-4 d-flex justify-content-end">
-                            <span className="fw-semibold" style={{ fontSize: '1.08rem', color: '#222', marginRight: 6 }}>작성자:</span>
-                            <span className="fw-bold" style={{ fontSize: '1.08rem', minWidth: 80, display: 'inline-block' }}>
-                                {post.memberNickname || ""}
-                            </span>
+                    ))}
+                    {newImages.map((file, idx) => (
+                        <div key={`new-${idx}`} style={{ position: 'relative' }}>
+                            <img
+                                src={imagePreviews[existingImages.length + idx]}
+                                alt={`preview-new-${idx}`}
+                                style={{ width: 100, height: 100, objectFit: 'cover', borderRadius: 14, border: '1px solid #eee', boxShadow: "0 2px 6px rgba(0,0,0,0.06)" }}
+                            />
+                            <button
+                                type="button" className="btn btn-danger btn-sm"
+                                style={{ position: 'absolute', top: 5, right: 5, borderRadius: '50%', padding: '2px 7px', fontSize: "1.05rem" }}
+                                onClick={() => handleNewImageRemove(idx)}
+                            >×</button>
                         </div>
-                        <form onSubmit={handleSubmit}>
-                            <div className="mb-4">
-                                <label htmlFor="title" className="form-label fw-semibold">제목</label>
-                                <input type="text" className="form-control form-control-lg" id="title"
-                                    required maxLength={40} placeholder="제목을 입력하세요"
-                                    value={post.title} onChange={handleChange} />
-                            </div>
-                            <div className="row g-3 mb-4">
-                                <div className="col-md-5">
-                                    <label htmlFor="travelPlace" className="form-label fw-semibold">여행지 이름</label>
-                                    <input type="text" className="form-control" id="travelPlace"
-                                        required placeholder="예: 남산타워"
-                                        value={post.travelPlace} onChange={handleChange} />
-                                </div>
-                                <div className="col-md-7">
-                                    <label htmlFor="address" className="form-label fw-semibold">여행지 주소</label>
-                                    <input type="text" className="form-control" id="address"
-                                        required placeholder="예: 서울특별시 중구 남산공원길 105"
-                                        value={post.address} onChange={handleChange} />
-                                </div>
-                            </div>
-                            <div className="row g-3 mb-4">
-                                <div className="col-md-6">
-                                    <label className="form-label fw-semibold">카테고리</label>
-                                    <div className="bg-light rounded-4 p-2 px-3 border">
-                                        <CategoryCard selectedCategory={post.category || ''} setCategory={handleCategorySelect} />
-                                    </div>
-                                </div>
-                                <div className="col-md-6">
-                                    <label className="form-label fw-semibold">지역 선택</label>
-                                    <div className="bg-light rounded-4 p-2 px-3 border">
-                                        <RadioPage selectedRegion={post.region} setRegion={handleRegionChange} />
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="mb-4">
-                                <label className="form-label fw-semibold">사진 첨부 <span className="text-secondary" style={{ fontSize: "0.95em" }}>(여러 장 첨부 가능)</span></label>
-                                <div className="bg-light rounded-4 p-3 px-4 border">
-                                    <input type="file" accept="image/*" multiple onChange={handleImageChange} className="form-control mb-3" />
-                                    <div className="d-flex flex-wrap gap-3">
-                                        {existingImages.map((src, idx) => (
-                                            <div key={`exist-${idx}`} style={{ position: 'relative' }}>
-                                                <img
-                                                    src={src.startsWith('/images/') ? `${process.env.REACT_APP_IMAGE_BASE_URL}${src}` : src}
-                                                    alt={`preview-exist-${idx}`}
-                                                    style={{ width: 100, height: 100, objectFit: 'cover', borderRadius: 14, border: '1px solid #eee', boxShadow: "0 2px 6px rgba(0,0,0,0.06)" }}
-                                                />
-                                                <button type="button" className="btn btn-danger btn-sm"
-                                                    style={{ position: 'absolute', top: 5, right: 5, borderRadius: '50%', padding: '2px 7px', fontSize: "1.05rem" }}
-                                                    onClick={() => handleExistingImageRemove(idx)}>×</button>
-                                            </div>
-                                        ))}
-                                        {newImages.map((file, idx) => (
-                                            <div key={`new-${idx}`} style={{ position: 'relative' }}>
-                                                <img src={imagePreviews[existingImages.length + idx]}
-                                                    alt={`preview-new-${idx}`}
-                                                    style={{ width: 100, height: 100, objectFit: 'cover', borderRadius: 14, border: '1px solid #eee', boxShadow: "0 2px 6px rgba(0,0,0,0.06)" }}
-                                                />
-                                                <button type="button" className="btn btn-danger btn-sm"
-                                                    style={{ position: 'absolute', top: 5, right: 5, borderRadius: '50%', padding: '2px 7px', fontSize: "1.05rem" }}
-                                                    onClick={() => handleNewImageRemove(idx)}>×</button>
-                                            </div>
-                                        ))}
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="mb-4">
-                                <label htmlFor="content" className="form-label fw-semibold">후기 / 내용</label>
-                                <textarea className="form-control" id="content" rows="7"
-                                    required maxLength={2000}
-                                    placeholder="여행지에 대한 후기를 자유롭게 작성해 주세요 :)"
-                                    value={post.content} onChange={handleChange} style={{ minHeight: 140 }}></textarea>
-                            </div>
-                            <div className="d-flex justify-content-between pt-2">
-                                <button className="btn btn-danger px-5 py-2 fs-5 fw-bold rounded-pill shadow"
-                                    type="button" style={{ minWidth: 140, letterSpacing: '1px' }}
-                                    onClick={removeBoard}>
-                                    삭제
-                                </button>
-                                <button type="submit"
-                                    className="btn btn-primary px-5 py-2 fs-5 fw-bold rounded-pill shadow"
-                                    style={{ minWidth: 140, letterSpacing: '1px' }}>
-                                    완료
-                                </button>
-                            </div>
-                        </form>
-                    </div>
+                    ))}
                 </div>
             </div>
-        </>
+        </div>
+    );
+
+    return (
+        <PostForm
+            post={post}
+            mode="edit"
+            onChange={handleChange}
+            onCategorySelect={handleCategorySelect}
+            onRegionChange={handleRegionChange}
+            onSubmit={handleSubmit}
+            onDelete={removeBoard}
+            imageSection={imageSection}
+            alert={alert}
+        />
     );
 };
